+++ packages/core/src/app/checkout/Checkout.spec.tsx
@@ -13,6 +13,7 @@ import { act } from 'react-dom/test-utils';
 
 import { AnalyticsContextProps, AnalyticsEvents, AnalyticsProviderMock } from '@bigcommerce/checkout/analytics';
 import { getLanguageService, LocaleProvider } from '@bigcommerce/checkout/locale';
+import { CheckoutProvider } from '@bigcommerce/checkout/payment-integration-api';
 
 import { BillingProps } from '../billing';
 import Billing from '../billing/Billing';
@@ -36,7 +37,6 @@ import Shipping from '../shipping/Shipping';
 import { getShippingAddress } from '../shipping/shipping-addresses.mock';
 
 import Checkout, { CheckoutProps, WithCheckoutProps } from './Checkout';
-import CheckoutProvider from './CheckoutProvider';
 import { getCheckout, getCheckoutWithPromotions } from './checkouts.mock';
 import CheckoutStep, { CheckoutStepProps } from './CheckoutStep';
 import CheckoutStepType from './CheckoutStepType';