+++ packages/core/src/app/customer/CheckoutButtonList.tsx
@@ -32,6 +32,7 @@ export const SUPPORTED_METHODS: string[] = [
     'googlepayorbital',
     'googlepaystripe',
     'googlepaystripeupe',
+    'googlepayworldpayaccess',
 ];
 
 export interface CheckoutButtonListProps {