+++ packages/core/src/app/customer/checkoutSuggestion/CheckoutSuggestion.tsx
@@ -7,9 +7,11 @@ import {
 import React, { FunctionComponent, memo } from 'react';
 
 import { useAnalytics } from '@bigcommerce/checkout/analytics';
+import { CheckoutContextProps } from '@bigcommerce/checkout/payment-integration-api';
 
-import { CheckoutContextProps, withCheckout } from '../../checkout';
+import { withCheckout } from '../../checkout';
 import { PaymentMethodId } from '../../payment/paymentMethod';
+
 
 import BoltCheckoutSuggestion from './BoltCheckoutSuggestion';