+++ packages/core/src/app/payment/SpamProtectionField.tsx
@@ -3,8 +3,9 @@ import { noop } from 'lodash';
 import React, { Component, MouseEvent, ReactNode } from 'react';
 
 import { TranslatedString } from '@bigcommerce/checkout/locale';
+import { CheckoutContextProps } from '@bigcommerce/checkout/payment-integration-api';
 
-import { CheckoutContextProps, withCheckout } from '../checkout';
+import { withCheckout } from '../checkout';
 import { isErrorWithType } from '../common/error';
 import { LoadingOverlay } from '../ui/loading';