+++ packages/core/src/app/payment/paymentMethod/CreditCardPaymentMethod.spec.tsx
@@ -11,9 +11,9 @@ import React, { FunctionComponent } from 'react';
 import { object, Schema, string } from 'yup';
 
 import { createLocaleContext, LocaleContext, LocaleContextType } from '@bigcommerce/checkout/locale';
+import { CheckoutProvider } from '@bigcommerce/checkout/payment-integration-api';
 
 import { getCart } from '../../cart/carts.mock';
-import { CheckoutProvider } from '../../checkout';
 import { getStoreConfig } from '../../config/config.mock';
 import { getCustomer } from '../../customer/customers.mock';
 import { getConsignment } from '../../shipping/consignment.mock';