+++ packages/core/src/app/payment/paymentMethod/PaymentMethodV2.tsx
@@ -3,6 +3,7 @@ import React, { ComponentType } from 'react';
 
 import { withLanguage, WithLanguageProps } from '@bigcommerce/checkout/locale';
 import {
+    PaymentFormProvider,
     PaymentFormValues,
     PaymentMethodResolveId,
     PaymentMethodProps as ResolvedPaymentMethodProps,
@@ -95,15 +96,19 @@ const PaymentMethodContainer: ComponentType<
         );
     }
 
+    const paymentForm = createPaymentFormService(formikContext, formContext, paymentContext);
+
     return (
-        <ResolvedPaymentMethod
-            checkoutService={checkoutService}
-            checkoutState={checkoutState}
-            language={language}
-            method={method}
-            onUnhandledError={onUnhandledError}
-            paymentForm={createPaymentFormService(formikContext, formContext, paymentContext)}
-        />
+        <PaymentFormProvider paymentForm={paymentForm}>
+            <ResolvedPaymentMethod
+                checkoutService={checkoutService}
+                checkoutState={checkoutState}
+                language={language}
+                method={method}
+                onUnhandledError={onUnhandledError}
+                paymentForm={paymentForm}
+            />
+        </PaymentFormProvider>
     );
 };