+++ packages/core/src/app/payment/storedInstrument/ManageInstrumentsModal.tsx
@@ -3,8 +3,10 @@ import { noop } from 'lodash';
 import React, { Component, ReactNode } from 'react';
 
 import { TranslatedString } from '@bigcommerce/checkout/locale';
+import { CheckoutContextProps } from '@bigcommerce/checkout/payment-integration-api';
 
-import { CheckoutContextProps, withCheckout } from '../../checkout';
+
+import { withCheckout } from '../../checkout';
 import { Button, ButtonSize, ButtonVariant } from '../../ui/button';
 import { Modal, ModalHeader } from '../../ui/modal';