+++ packages/payment-integration-api/src/contexts/checkout-context/index.ts
@@ -1,1 +1,2 @@
-export { CheckoutContext, CheckoutContextProps, useCheckout } from './CheckoutContext';
+export { default as CheckoutContext, CheckoutContextProps, useCheckout } from './CheckoutContext';
+export { default as CheckoutProvider } from './CheckoutProvider';